+++ app/models/delivery_param.rb
@@ -1,0 +1,4 @@
+# Represents delivery params for deliveries
+class DeliveryParam < ApplicationRecord
+  belongs_to :delivery
+end