+++ spec/factories/delivery_params.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :delivery_param do
+    params { { 'id' => '1', 'token' => '1' } }
+    association :delivery
+  end
+end