+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - discord_bot library
 - Notification model
 - Delivery model
+- DeliveryParam model
 
 ## [0.6.2] - 2019-11-04
 ### Added